+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.16??
 - Fix Feature initialization from GeoJSON: respect configured `geomField`, apply optional non-standard embedded `srid` and `id` correctly
+- Fix broken data format in Oracle::prepareResults
 - Support `:userName` filter binding also in DataStore::search (previously only in FeatureType::search)
 - Escape `:userName` properly in FeatureType::search and DataStore::search
 - Escape `:distance` in FeatureType::search (now a bound param)